Made product price accept float type

+++ database/migrations/2022_09_23_084425_create_date_produses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->string('nume');
             $table->string('um');
-            $table->integer('pret');
+            $table->float('pret');
             $table->string('moneda');
             $table->integer('cota_tva');
             $table->boolean('tva');
